Extract annotation related functions / hooks into separate file
