+++ src/patchwork/utils.ts
@@ -512,65 +512,3 @@ export function useDebounce<T>(value: T, delay?: number): T {
 
   return debouncedValue;
 }
-
-export const doAnchorsOverlap = (
-  type: DocType,
-  a: unknown,
-  b: unknown,
-  doc: HasPatchworkMetadata<unknown, unknown>
-) => {
-  const comperator = docTypes[type].doAnchorsOverlap;
-  return comperator ? comperator(a, b, doc) : isEqual(a, b);
-};
-
-export const areAnchorSelectionsEqual = (
-  type: DocType,
-  a: unknown[],
-  b: unknown[],
-  doc
-) => {
-  if (a.length !== b.length) {
-    return false;
-  }
-
-  return a.every((anchor) =>
-    b.some((other) => doAnchorsOverlap(type, anchor, other, doc))
-  );
-};
-
-export function getAnnotationGroupId<T, V>(
-  annotationGroup: AnnotationGroup<T, V>
-) {
-  if (annotationGroup.discussion) return annotationGroup.discussion.id;
-
-  // if the annotation group has no discussion we know that it's a computed annotation group
-  // which means that the annotation doesn't appear in any other annotationGroup
-  // so we can just pick the first annotation to generate a unique id
-  const firstAnnotation = annotationGroup.annotations[0];
-  return `${firstAnnotation.type}:${JSON.stringify(firstAnnotation.target)}`;
-}
-
-export function doesAnnotationGroupContainAnchors<T, V>(
-  docType: DocType,
-  group: AnnotationGroup<T, V>,
-  anchors: T[],
-  doc: HasPatchworkMetadata<T, V>
-) {
-  return anchors.every((anchor) =>
-    group.annotations.some((annotation) =>
-      doAnchorsOverlap(docType, annotation.target, anchor, doc)
-    )
-  );
-}
-
-export function groupAnnotations<T, V>(
-  docType: DocType,
-  annotations: Annotation<T, V>[]
-): Annotation<T, V>[][] {
-  const grouper =
-    docTypes[docType].groupAnnotations ??
-    ((annotations: Annotation<T, V>[]) =>
-      annotations.map((annotation) => [annotation]));
-
-  return grouper(annotations) as Annotation<T, V>[][];
-}
